Rename File and extract method
